EnumGen added -Seq start step- enumeration scheme
Now you can use, besides Manual and Auto,
a sequential numbering scheme which allows
you to define a sequence with a start and a step
between each element. First parameter after Seq
is the starting value and the second is the step.
If step is not given, it defaults to 1 and if both
parameters are missing, start will also default to 1.
It disregards any parameter after the second one.

The template generator now adds a function
in the form of to<your-enum-name> to make it
more convenient to find your enum item from the
int id, in case you have lots of enums.